Move platforms to paths
