--- apps/robots-generator/src/components/PlatformSettings/index.ts
@@ -1,3 +1,0 @@
-import PlatformSettings from "./PlatformSettings";
-
-export default PlatformSettings;
